tasks/requirements.txt
luigi
numpy
pandas
h5py
PyYAML==3.12
beautifulsoup4==4.5.3
elasticsearch==5.1.0
certifi==2016.9.26
coreapi==2.1.1
coreapi-cli==1.0.6

